requirement.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pytesseract==0.3.10
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
